Smarter job building
Don't just build when we hit a leaf we can't add. Instead, put as many packages
into each job as we can. Defer a packages if another version is already in the
job, or if a different version of its direct dependencies is already in the job.
If we loop through all current leaf packages without adding anything, or the
number of packages in the job goes over the chunk size, then build what we
have.