Make package cache backend work (VAPPPLAT-1372)
The package cache backend was not fully fleshed out. This patch makes
the package cache backend operational and works around an inconsistency
in the package cache data and what was originally imported into the
conary repository. It looks like some of the original imports included
full paths to source rpms in their manifests rather than leveraging
synthesized sources. The package cache was only populated with
synthesized sources. Thus, work was required to map the synthesized
paths and the real paths together.